+++ configs/small-scale-node-density-cni.yaml
@@ -1,4 +1,6 @@
 # This is a template file
+datasource:
+  type: perfscale
 tests :
   - name : aws-small-scale-node-density-cni
     index: ospst-perf-scale-ci-*
