+++ examples/payload-scale-416.yaml
@@ -1,3 +1,5 @@
+datasource:
+  type: perfscale
 tests :
   - name : aws-small-scale-cluster-density-v2
     index: {{ es_metadata_index }}
