+++ examples/readout-control-plane-cdv2.yaml
@@ -1,3 +1,5 @@
+datasource:
+  type: perfscale
 tests :
   - name : cluster-density-v2-24nodes
     index: {{ es_metadata_index }}
