+++ examples/readout-control-plane-node-density.yaml
@@ -1,3 +1,5 @@
+datasource:
+  type: perfscale
 tests :
   - name : node-density-heavy-24nodes
     index: {{ es_metadata_index }}
