+++ examples/readout-netperf-tcp.yaml
@@ -1,3 +1,5 @@
+datasource:
+  type: perfscale
 tests :
   - name : k8s-netperf-tcp
     index: {{ es_metadata_index }}
