+++ examples/small-scale-node-density-cni.yaml
@@ -1,3 +1,5 @@
+datasource:
+  type: perfscale
 tests :
   - name : aws-small-scale-node-density-cni
     index: {{ es_metadata_index }}
