+++ examples/trt-external-payload-cluster-density.yaml
@@ -1,3 +1,5 @@
+datasource:
+  type: perfscale
 tests :
   - name : payload-cluster-density-v2
     index: {{ es_metadata_index }}
