+++ output_aws-small-scale-cluster-density-v2.txt
@@ -1,26 +1,24 @@
 time                       uuid                                    total_cpu    os_daemon
 -------------------------  ------------------------------------  -----------  -----------
-2024-09-01 09:45:59 +0000  14da4d9c-b0f5-4230-98c0-d9e096508dd0     0.511546    0.0946667
-2024-09-02 10:42:47 +0000  4bb33108-8ed0-4ae6-9b49-9fde0c0b4fd1     0.579733    0.178
-2024-09-04 16:24:34 +0000  637a8804-333c-4976-b4cf-5423702a2a61     0.673504    0.194667
-2024-09-08 09:33:58 +0000  d810faf2-2c93-49be-b6d8-20f3201ca58f     0.484272    0.088
-2024-09-08 11:01:36 +0000  ee43117d-c710-4657-865b-c9af14d876fa     0.586345    0.121333
-2024-09-09 10:50:10 +0000  0870b8fa-5f22-4d4f-bde2-4f42043a78c0     0.572865    0.138
-2024-09-12 14:05:24 +0000  d85c2bc1-328b-42c4-8fc5-c6f3fcc9ff7f     0.572865    0.138
-2024-09-12 14:06:53 +0000  40e83e58-aaa1-4ec2-bfaa-da90dd9a8ed0     0.572865    0.138
-2024-09-12 14:08:39 +0000  53155341-89c6-4f20-a0d2-e10221b351f6     0.572865    0.138
-2024-09-12 14:12:48 +0000  e98e01b1-a426-4935-bb55-7e3392c00778     0.572865    0.138
-2024-09-12 14:32:23 +0000  fb52cec9-cb44-4e61-91b5-d7aa073c6ce1     0.572865    0.138
-2024-09-12 15:06:57 +0000  7ccd05b7-74df-47df-980f-6f4fbc8b7e1c     0.572865    0.138
-2024-09-12 15:11:31 +0000  32edfa2f-5965-4974-afd4-6cf44b2401a6     0.572865    0.138
-2024-09-12 16:50:53 +0000  6c493ed2-5eba-4ebc-986f-89da35974ca3     0.572865    0.138
-2024-09-12 16:53:54 +0000  9fddd8d3-55a7-4a90-bc5f-e33d329fa74b     0.572865    0.138
-2024-09-12 21:20:30 +0000  967cc3cc-d8d1-407d-a995-ee6f6500ed3b     0.594241    0.146667
-2024-09-15 09:41:52 +0000  ccdd9a09-d68b-41e3-82fa-b5e927e149db     0.481645    0.0913333
-2024-09-15 11:03:26 +0000  fc38367b-3951-4b5f-9579-07a6177efad5     0.640491    0.161333
-2024-09-16 11:01:18 +0000  9abd0b68-f0ee-41ab-b4c8-b7aa75726add     0.555351    0.159333
-2024-09-24 03:56:58 +0000  be47580e-1a72-4e03-a509-7f918568f1b5     0.597396    0.121333
-2024-09-28 05:20:42 +0000  0c49f237-7cc7-41c9-b58c-13efbf962a55     0.567658    0.094
-2024-09-29 10:03:24 +0000  ce12a265-0b69-4bc3-92cc-2772c6bb0171     0.554108    0.144
-2024-09-29 11:21:35 +0000  cfadb284-22b8-4673-a918-8aa75c57c2ef     0.593514    0.132
-2024-09-30 11:10:27 +0000  1373cc13-535f-4aa2-ac31-40d64d66fa99     0.625041    0.195333
+2024-09-08 09:33:58 +0000  ed1739a9-bf98-4e1d-ac91-91f48bbbc2ac     0.484272    0.088
+2024-09-08 11:01:36 +0000  41c478e3-b356-45f6-a46b-9e8d0fe5e3dd     0.586345    0.121333
+2024-09-09 10:50:10 +0000  2174a2cd-aa2a-4177-a55f-065bd8974681     0.572865    0.138
+2024-09-12 14:05:24 +0000  26673741-44e2-4f42-acbf-21a3e306a500     0.572865    0.138
+2024-09-12 14:06:53 +0000  575cebdd-fb95-4d84-b6fa-59a35841a546     0.572865    0.138
+2024-09-12 14:08:39 +0000  82f74ece-e20b-4710-89d2-6062cd9e962c     0.572865    0.138
+2024-09-12 14:12:48 +0000  ae1e8a9d-ee0a-4e58-a019-76b50942666f     0.572865    0.138
+2024-09-12 14:32:23 +0000  33a8223b-4364-4fbe-8b32-81baebaa8b47     0.572865    0.138
+2024-09-12 15:06:57 +0000  44bb2eac-0ee6-48c5-a8e3-42ceffd5fe27     0.572865    0.138
+2024-09-12 15:11:31 +0000  38f78f63-6754-4565-96f6-4c552ee568fb     0.572865    0.138
+2024-09-12 16:50:53 +0000  3a155cfb-14f2-4d62-8799-78a826f53791     0.572865    0.138
+2024-09-12 16:53:54 +0000  85e8a2d7-be99-49b4-91a2-2a09cd5ab2c2     0.572865    0.138
+2024-09-12 21:20:30 +0000  7243b987-be5d-49e9-9f02-c6cd156043f8     0.594241    0.146667
+2024-09-15 09:41:52 +0000  b943dc95-14cb-4375-9b69-49123e1da97a     0.481645    0.0913333
+2024-09-15 11:03:26 +0000  3bb70869-eac6-49b8-99c7-8a409994d7e7     0.640491    0.161333
+2024-09-16 11:01:18 +0000  7a221281-8056-4deb-a91c-c52ede9f67ff     0.555351    0.159333
+2024-09-24 03:56:58 +0000  3cfefcd0-c79f-4b1c-b2f7-101e49031f00     0.597396    0.121333
+2024-09-28 05:20:42 +0000  d40001f3-7f98-4fcf-838c-94e440dce9c1     0.567658    0.094
+2024-09-29 10:03:24 +0000  3a7673b1-ecff-493f-b763-445261997ad2     0.554108    0.144
+2024-09-29 11:21:35 +0000  f4ca17d2-a87e-4d2a-9c1e-3d635190bc7a     0.593514    0.132
+2024-09-30 11:10:27 +0000  a0d4865c-4c56-4e95-bf9a-9d2600566f1d     0.625041    0.195333
+2024-10-01 12:02:40 +0000  8703bb88-5c3c-4caf-8958-a607bba6cf77     0.652305    0.122667
