+++ output_telco_cpu_util_test.txt
@@ -1,0 +1,28 @@
+<?xml version="1.0" ?>
+<testsuites>
+	<testsuite name="telco_cpu_util_test nightly compare" failures="1" tests="2">
+		<testcase name=" total_cpu regression detection" timestamp="1728336710">
+			<failure>
++----+--------------------------------------+----------------------+------------------------------+-------------+------------------+---------------------+
+|    | uuid                                 | timestamp            | buildUrl                     |   total_cpu | is_changepoint   |   percentage_change |
+|----+--------------------------------------+----------------------+------------------------------+-------------+------------------+---------------------|
+|  0 | ff28201c-3b73-c65e-2169-0645c9d68c2f | 2024-09-08T09:33:58Z | https://placeholder.com/1350 |    0.484272 | False            |              0      |
+|  1 | 2558dbab-ec77-ba3d-32b3-9306d256e8ea | 2024-09-12T14:05:24Z | https://placeholder.com/1357 |    0.572865 | True             |             18.7354 | -- changepoint
+|  2 | f4484f00-a13c-41ca-9119-381260062e6f | 2024-09-12T14:06:53Z | https://placeholder.com/1357 |    0.572865 | False            |              0      |
+|  3 | 8b9c7ba6-8447-65b3-89ca-a6842cc8d5e7 | 2024-09-12T14:08:39Z | https://placeholder.com/1357 |    0.572865 | False            |              0      |
+|  4 | f0065f63-2543-de82-1818-8b25b70c2c29 | 2024-09-12T14:12:48Z | https://placeholder.com/1357 |    0.572865 | False            |              0      |
+|  5 | 56814eb4-07ec-06a8-6227-1480f882e258 | 2024-09-12T14:32:23Z | https://placeholder.com/1357 |    0.572865 | False            |              0      |
+|  6 | 179004d2-de86-5920-a30f-02f4d96642a9 | 2024-09-12T15:06:57Z | https://placeholder.com/1357 |    0.572865 | False            |              0      |
+|  7 | 7ac54f43-4852-74dc-998d-9f2a693cf0a6 | 2024-09-12T15:11:31Z | https://placeholder.com/1357 |    0.572865 | False            |              0      |
+|  8 | 42f90750-025d-8664-f9d8-7aa4dd9e0185 | 2024-09-12T16:50:53Z | https://placeholder.com/1357 |    0.572865 | False            |              0      |
+|  9 | 1a377473-94e8-6fa0-7b7f-6dadd36c5164 | 2024-09-12T16:53:54Z | https://placeholder.com/1357 |    0.572865 | False            |              0      |
+| 10 | 7bc9ba32-413e-7d61-c2e4-88a7046a8bf9 | 2024-09-12T21:20:30Z | https://placeholder.com/1368 |    0.594241 | False            |              0      |
+| 11 | 1d366328-34b8-059d-cec6-40778366dc69 | 2024-09-15T09:41:52Z | https://placeholder.com/1396 |    0.481645 | False            |              0      |
+| 12 | 8ad29635-c8b9-32d1-110b-c6968790c008 | 2024-09-28T05:20:42Z | https://placeholder.com/287  |    0.567658 | False            |              0      |
+| 13 | a573ad1e-0f04-097b-724f-7312f2e6e1d8 | 2024-09-29T10:03:24Z | https://placeholder.com/1479 |    0.554108 | False            |              0      |
++----+--------------------------------------+----------------------+------------------------------+-------------+------------------+---------------------+
+</failure>
+		</testcase>
+		<testcase name=" os_daemon regression detection" timestamp="1728336710"/>
+	</testsuite>
+</testsuites>
