+++ pkg/algorithms/edivisive/edivisive.py
@@ -27,8 +27,8 @@ class EDivisive(Algorithm):
         # filter by direction
         for metric, changepoint_list in change_points_by_metric.items():
             for i in range(len(changepoint_list)-1, -1, -1):
-                if ((self.metrics_config[metric]["direction"] == 1 and changepoint_list[i].stats.mean_1 > changepoint_list[i].stats.mean_2) or
-                    (self.metrics_config[metric]["direction"] == -1 and changepoint_list[i].stats.mean_1 < changepoint_list[i].stats.mean_2) ):
+                if ((self.metrics_config[metric].get("direction",0) == 1 and changepoint_list[i].stats.mean_1 > changepoint_list[i].stats.mean_2) or
+                    (self.metrics_config[metric].get("direction",0) == -1 and changepoint_list[i].stats.mean_1 < changepoint_list[i].stats.mean_2) ):
                     del changepoint_list[i]
         if [val for li in change_points_by_metric.values() for val in li]:
             self.regression_flag=True
