+++ pkg/datasources/__init__.py
@@ -1,0 +1,7 @@
+"""
+Datasource module for orion
+"""
+from .datasource import Datasource
+from .datasourceFactory import DatasourceFactory
+from .perfscale import PerfscaleDatasource
+from .telco import TelcoDatasource
