+++ pkg/datasources/perfscale/__init__.py
@@ -1,0 +1,4 @@
+"""
+Perfscale datasource init module
+"""
+from .perfscale import PerfscaleDatasource
