+++ pkg/runTest.py
@@ -1,14 +1,16 @@
 """
 run test
 """
+
+import asyncio
 import sys
 from typing import Any, Dict
-from fmatch.matcher import Matcher
-from fmatch.splunk_matcher import SplunkMatcher
 from fmatch.logrus import SingletonLogger
 from pkg.algorithms import AlgorithmFactory
 import pkg.constants as cnsts
-from pkg.utils import get_datasource, process_test, get_subtracted_timestamp, process_splunk_test
+from pkg.utils import get_datasource, get_subtracted_timestamp
+from pkg.datasources import DatasourceFactory
+
 
 def get_algorithm_type(kwargs):
     """Switch Case of getting algorithm name
@@ -23,13 +25,14 @@ def get_algorithm_type(kwargs):
         algorithm_name = cnsts.EDIVISIVE
     elif kwargs["anomaly_detection"]:
         algorithm_name = cnsts.ISOLATION_FOREST
-    elif kwargs['cmr']:
+    elif kwargs["cmr"]:
         algorithm_name = cnsts.CMR
     else:
         algorithm_name = None
     return algorithm_name
 
-async def run(**kwargs: dict[str, Any]) -> dict[str, Any]: #pylint: disable = R0914
+
+async def run(**kwargs: dict[str, Any]) -> dict[str, Any]:  # pylint: disable = R0914
     """run method to start the tests
 
     Args:
@@ -49,38 +52,26 @@ async def run(**kwargs: dict[str, Any]) -> dict[str, Any]: #pylint: disable = R0
     fingerprint_matched_df, metrics_config = None, None
     for test in config_map["tests"]:
         # Create fingerprint Matcher
-        if datasource["type"] == "elasticsearch":
-            matcher = Matcher(
-                index=test["index"],
-                level=logger_instance.level,
-                ES_URL=datasource["ES_SERVER"],
-                verify_certs=False,
+        start_timestamp = get_start_timestamp(kwargs)
+        datasourceFactory = DatasourceFactory()
+        datasource_object, matcher = datasourceFactory.instantiate_datasource(
+            datasource=datasource,
+            test=test,
+            options=kwargs,
+            start_timestamp=start_timestamp,
+        )
+        if asyncio.iscoroutinefunction(datasource_object.process_test):
+            fingerprint_matched_df, metrics_config = (
+                await datasource_object.process_test()
             )
-            start_timestamp = get_start_timestamp(kwargs)
-            fingerprint_matched_df, metrics_config = process_test(
-                test,
-                matcher,
-                kwargs,
-                start_timestamp,
-            )
-        elif datasource["type"] == "splunk":
-            matcher = SplunkMatcher(
-                host= datasource.get("host"),
-                port= datasource.get("port"),
-                username= datasource.get("username"),
-                password= datasource.get("password"),
-                indice=datasource.get("indice")
-            )
-            start_timestamp = get_start_timestamp(kwargs)
-            fingerprint_matched_df, metrics_config = await process_splunk_test(
-                test,
-                matcher,
-                kwargs,
-                start_timestamp,
-            )
+        else:
+            fingerprint_matched_df, metrics_config = datasource_object.process_test()
 
         if fingerprint_matched_df is None:
-            sys.exit(3) # No data present
+            sys.exit(3)  # No data present
+        logger_instance.debug(
+            f"Collected dataframe {fingerprint_matched_df},\n metrics {metrics_config}"
+        )
 
         algorithm_name = get_algorithm_type(kwargs)
         if algorithm_name is None:
@@ -88,14 +79,15 @@ async def run(**kwargs: dict[str, Any]) -> dict[str, Any]: #pylint: disable = R0
 
         algorithmFactory = AlgorithmFactory()
         algorithm = algorithmFactory.instantiate_algorithm(
-                algorithm_name,
-                matcher,
-                fingerprint_matched_df,
-                test,
-                kwargs,
-                metrics_config,
-            )
+            algorithm_name,
+            matcher,
+            fingerprint_matched_df,
+            test,
+            kwargs,
+            metrics_config,
+        )
         testname, result_data, test_flag = algorithm.output(kwargs["output_format"])
+        logger_instance.debug(f"Result data for test {testname}, {result_data}")
         result_output[testname] = result_data
         regression_flag = regression_flag or test_flag
     return result_output, regression_flag
